feat(search): offload AI search to worker thread
